web25-login-css: moved auth work to own compnent folder and updated links, removed commemtns out code
