+++ wetmap/src/components/authenticationPage.jsx
@@ -1,15 +1,11 @@
-import { useState } from "react";
-import PropTypes from "prop-types";
-import Tabs from "@mui/material/Tabs";
-import Tab from "@mui/material/Tab";
-import Typography from "@mui/material/Typography";
-import Box from "@mui/material/Box";
-import SignInRoute from "./signIn";
-import SignUpRoute from "./signUp";
-
-import Button from "./reusables/button";
-
-import AuthticationPage from "./pages/authentication";
+import { useState } from 'react';
+import PropTypes from 'prop-types';
+import Tabs from '@mui/material/Tabs';
+import Tab from '@mui/material/Tab';
+import Box from '@mui/material/Box';
+import SignInRoute from './signIn';
+import SignUpRoute from './signUp';
+import AuthticationPage from './authentication';
 
 function TabPanel(props) {
   const { children, value, index, ...other } = props;
@@ -33,14 +29,14 @@ function TabPanel(props) {
 
 TabPanel.propTypes = {
   children: PropTypes.node,
-  index: PropTypes.number.isRequired,
-  value: PropTypes.number.isRequired,
+  index:    PropTypes.number.isRequired,
+  value:    PropTypes.number.isRequired,
 };
 
 function a11yProps(index) {
   return {
-    id: `simple-tab-${index}`,
-    "aria-controls": `simple-tabpanel-${index}`,
+    'id':              `simple-tab-${index}`,
+    'aria-controls': `simple-tabpanel-${index}`,
   };
 }
 
@@ -55,38 +51,38 @@ export default function AuthenticationPage() {
     <>
       <AuthticationPage />
 
-      <Box sx={{ width: "100vw", height: "100vh" }}>
-        <Box sx={{ borderBottom: 1, borderColor: "lightgrey", width: "100vw" }}>
+      <Box sx={{ width: '100vw', height: '100vh' }}>
+        <Box sx={{ borderBottom: 1, borderColor: 'lightgrey', width: '100vw' }}>
           <Tabs
             value={value}
             onChange={handleChange}
             TabIndicatorProps={{
-              sx: { backgroundColor: "lightgray", height: 2 },
+              sx: { backgroundColor: 'lightgray', height: 2 },
             }}
             sx={{
-              "& button.Mui-selected": { color: "lightgray", width: "50%" },
-              backgroundColor: "#538dbd",
-              fontFamily: "Patrick Hand",
+              '& button.Mui-selected': { color: 'lightgray', width: '50%' },
+              'backgroundColor':         '#538dbd',
+              'fontFamily':              'Patrick Hand',
             }}
           >
             <Tab
               label="Sign In"
               {...a11yProps(0)}
               sx={{
-                fontFamily: "Patrick Hand",
-                color: "darkgray",
-                width: "50%",
-                fontSize: "2.5vw",
+                fontFamily: 'Patrick Hand',
+                color:      'darkgray',
+                width:      '50%',
+                fontSize:   '2.5vw',
               }}
             />
             <Tab
               label="Sign Up"
               {...a11yProps(1)}
               sx={{
-                fontFamily: "Patrick Hand",
-                color: "darkgray",
-                width: "50%",
-                fontSize: "2.5vw",
+                fontFamily: 'Patrick Hand',
+                color:      'darkgray',
+                width:      '50%',
+                fontSize:   '2.5vw',
               }}
             />
           </Tabs>
@@ -94,14 +90,14 @@ export default function AuthenticationPage() {
         <TabPanel
           value={value}
           index={0}
-          style={{ backgroundColor: "#538dbd", height: "100%" }}
+          style={{ backgroundColor: '#538dbd', height: '100%' }}
         >
           <SignInRoute setValue={setValue} />
         </TabPanel>
         <TabPanel
           value={value}
           index={1}
-          style={{ backgroundColor: "#538dbd", height: "100%" }}
+          style={{ backgroundColor: '#538dbd', height: '100%' }}
         >
           <SignUpRoute />
         </TabPanel>
